+++ src/main/java/grandduke/exception/event/EmptyEventException.java
@@ -1,4 +1,6 @@
-package grandduke.exception;
+package grandduke.exception.event;
+
+import grandduke.exception.GrandException;
 
 public class EmptyEventException extends GrandException {
     @Override
