+++ src/main/java/grandduke/exception/mark/MarkMissingDescriptionException.java
@@ -1,4 +1,6 @@
-package grandduke.exception;
+package grandduke.exception.mark;
+
+import grandduke.exception.GrandException;
 
 public class MarkMissingDescriptionException extends GrandException {
     @Override
